python/requirements.txt
#python 3.10
sbx-rl>=0.9.1
pyzmq
pathos>=0.3.1
pyarrow>=14.0.2

pytest>=7.4.4
pytest-cov>=4.1.0


cvxpy>=1.4.1
numpy
pandas>=2.1.4
numba>=0.58.1
tensorboard>=2.15.1
#onnxruntime
#onnxconverter-common
#onnx
#tf2onnx
#skl2onnx
#onnxmltools
#keras2onnx

#numerox
#numerapi

#cryptofeed
#cryptocompare
#python-binance
#alpha-vantage

git+https://github.com/oanda/oandapy.git
#mplfinance

boto3>=1.34.12
s3transfer>=0.10.0

tqdm>=4.66.1
yfinance>=0.2.33
empyrical>=0.5.5

dill>=0.3.7
nose-parameterized>=0.6.0

seaborn>=0.13.1
plotly>=5.18.0
statsmodels>=0.14.1

gymnasium>=0.29.1
#Shimmy
stable-baselines3>=2.2.1
scikit-learn>=1.3.2
#beautifulsoup4
dash>=2.14.2
dash_bootstrap_components>=1.5.0
dash_daq>=0.5.0
#markupsafe
pandas_datareader>=0.10.0
optuna>=3.5.0
skfolio>=0.0.10
